requirements.txt
langchain
openai
langchain_openai
streamlit
python-dotenv
langchain_community